client-hints/feature-policy-allow-cross-origin-ch.https.sub.html

<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>

promise_test(t => {
  
  // TODO do I need the port here?
  return fetch("https://{{hosts[][www]}}:{{ports[https][0]}}/client-hints/echo_client_hints_received.py").then(r => {
      assert_equals(r.status, 200);
      assert_true(r.headers.has("dpr"));
    })
  },
  "Enabled cross-origin Client Hints with Feature Policy; hint was sent." 
);

</script>

</body>
</html>
